Added saveOngoingRides, saveCompletedRides(String date), saveCompletedRides() for all date in dailyRidesMap.
The strings used for transforming String/Dates are not final fields in
inputUtil. There is a dateToString method that takes a string and a
desired format as parameters.

In ValleyBikeSim the dailyRidesMap is keyed by a String in the format
"yyyy-mm-dd".

Ongoing rides are saved after a rides is started. When a ride ends, the
file for that day is appended to. Completed rides are added to the
dailyRidesMap in endRide.

The bug in initialize was fixed.

Testing for equalize and adding rides to dailyRidesList was added.